wechat_automation/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
markdownify>=0.11.6
google-generativeai>=0.3.0
aiohttp>=3.9.0
pyyaml>=6.0.1
tqdm>=4.66.1
pillow>=10.0.0
lxml>=4.9.3
asyncio>=3.4.3
click>=8.1.7
colorama>=0.4.6